Verify byte values in KafkaIndexTaskTest
