+++ extensions-core/multi-stage-query/src/main/java/org/apache/druid/msq/input/external/ExternalInputSliceReader.java
@@ -26,7 +26,6 @@ import org.apache.druid.data.input.InputRow;
 import org.apache.druid.data.input.InputRowSchema;
 import org.apache.druid.data.input.InputSource;
 import org.apache.druid.data.input.InputSourceReader;
-import org.apache.druid.data.input.InputStats;
 import org.apache.druid.data.input.impl.DimensionsSpec;
 import org.apache.druid.data.input.impl.InlineInputSource;
 import org.apache.druid.data.input.impl.TimestampSpec;
@@ -53,7 +52,6 @@ import org.apache.druid.msq.util.DimensionSchemaUtils;
 import org.apache.druid.segment.RowAdapters;
 import org.apache.druid.segment.RowBasedSegment;
 import org.apache.druid.segment.column.RowSignature;
-import org.apache.druid.segment.incremental.SimpleRowIngestionMeters;
 import org.apache.druid.timeline.SegmentId;
 
 import java.io.File;
@@ -136,7 +134,6 @@ public class ExternalInputSliceReader implements InputSliceReader
     if (!temporaryDirectory.exists() && !temporaryDirectory.mkdir()) {
       throw new ISE("Cannot create temporary directory at [%s]", temporaryDirectory);
     }
-    final InputStats inputStats = new SimpleRowIngestionMeters();
     return Iterators.transform(
         inputSources.iterator(),
         inputSource -> {
@@ -162,7 +159,7 @@ public class ExternalInputSliceReader implements InputSliceReader
                     public CloseableIterator<InputRow> make()
                     {
                       try {
-                        CloseableIterator<InputRow> baseIterator = reader.read(inputStats);
+                        CloseableIterator<InputRow> baseIterator = reader.read();
                         return new CloseableIterator<InputRow>()
                         {
                           private InputRow next = null;
